# Add `stream` method for streaming content, add docs and examples